Mark storage version as part of pipeline
